Added Paginator
This has no relationship to any Paginator libraries
that may be available out there. This one is a rework
of the simple pagination code I started the other night.

This now formats nicely using Bootstrap, and works
reasonably well. Possibly some minor issues, but it does
get the point across.