Added Basic Search
This is a very rudimentary search feature that searches
the song form by artist. It could be expanded to search
on different fields if warranted.

It uses a GET instead of a POST so that we can more
easily handle the transition between pages. Seems
reasonable. Still some niggling issues including not
fully clearing the query string on the URL bar if the
search box is cleared.

SUPPORTING ADDITIONS:

* Two new queries to SongModel
* Updated the view with a very rudimentary form (not styled)
* Updated the controller to handle the search parameters
* Updated Paginator to handle the query string while in search mode

BUGFIXES:

* A couple of issues with Paginator with regard to low
  number record sets. I'm missing something about how to use
  the || (or) operator with PHP because it is not working
  how I would expect it to.